requirements.txt
# Telegram Bot Library
python-telegram-bot==21.0.1

# AI/ML APIs
voyageai==0.2.3
openai==1.54.3

# Database
psycopg2-binary==2.9.10

# Environment Variables
python-dotenv==1.0.1
